pyproject.toml
[tool.poetry]
name = "azure-sql"
version = "0.1.0"
description = ""
authors = ["Sergio León <panicoenlaxbox@gmail.com>"]
readme = "README.md"
packages = [{ include = "azure_sql", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.87.0" }
azure-mgmt-sql = "^3.0.1"
azure-mgmt-resource = "^21.2.1"
msrestazure = "^0.6.4"
azure-identity = "^1.12.0"
aiohttp = "^3.8.3"
azure-mgmt-subscription = "^3.1.1"
azure-keyvault-secrets = "^4.6.0"
asyncpg = "^0.27.0"
sqlalchemy = {extras = ["mypy"], version = "^1.4.44"}
alembic = "^1.8.1"
httpx = "^0.23.1"
dependency-injector = {extras = ["yaml"], version = "^4.40.0"}
mediatpy = "^0.2.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^5.0.4"
flake8-pyproject = "^1.2.0"
pytest = "^7.2.0"
assertpy = "^1.1"
pytest-asyncio = "^0.20.2"
pytest-cov = "^4.0.0"
mypy = "^0.991"
pre-commit = "^2.20.0"
python-dotenv = {extras = ["cli"], version = "^0.21.0"}
asyncpg-stubs = "^0.27.0"
psycopg2 = "^2.9.5"
types-psycopg2 = "^2.9.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.flake8]
max-line-length = 120
max-doc-length = 140

[tool.mypy]
show_error_codes = true
pretty = true
show_error_context = true
show_column_numbers = true
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true
# https://mypy.readthedocs.io/en/stable/config_file.html#confval-mypy_path
# https://mypy.readthedocs.io/en/stable/stubs.html#creating-a-stub
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"
plugins = [
    "sqlalchemy.ext.mypy.plugin"
]
exclude = [
    "migrations"
]

[[tool.mypy.overrides]]
module = [
    "assertpy",
    "azure.mgmt.sql.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = "src/"
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = ["pass"]
